calculate indices with negative thresholds

+++ index_calculator/_processing.py
@@ -92,12 +92,12 @@ class Processing:
             elif numb_name:
                 default_value = getattr(obj, attr)
                 if not isinstance(default_value, list):
-                    default_value = int(default_value)
+                    default_value = float(default_value)
                     if default_value < 0:
                         numb_name = "-{}".format(numb_name)
                 else:
                     continue
-                replacement[attr] = numb_name
+                replacement[attr] = float(numb_name)
             else:
                 replacement[attr] = getattr(obj, attr)
             if isinstance(replacement[attr], list):
@@ -114,6 +114,7 @@ class Processing:
                     )
                 else:
                     replacement[attr] = int(replacement[attr])
+
         return replacement
 
     def _get_idx_name_and_repl(self):
